requirements.txt
flask>=2.3.3
openai>=1.8.0
python-dotenv>=1.0.0
websocket-client>=1.6.4
websockets>=11.0.3
python-socketio>=5.9.0
flask-socketio>=5.3.5
gunicorn>=21.2.0
numpy>=1.26.0
soundfile>=0.12.1
setuptools>=68.0.0
wheel>=0.40.0
requests>=2.31.0
